FP comparisons do not use default constructor
Changed the occurences of FPT() to FPT(0). Since we expect a construction from int anyway,
and since the previous implementation assumed that the default constructor initializes to
numeric vaue zero, the change appears uncontroversial. What we gain, is that the default
ctor is no longer required. The FP algos work for non-default-constructible types.