Changed v == T(0)  to  v == 0
With this change we get a *desired* compile time failure when we are passed type `T` that is not implicitly convertible from `int` but is explicitly convertible. Now, we make an implicit conversion from `int` a requirement and refuse to compile otherwise.

This addresses an issue where the user gives us a type `T` with non-intuitive construction from `int`.